deps: update V8 to 10.7.193.13
PR-URL: https://github.com/nodejs/node/pull/44741
Fixes: https://github.com/nodejs/node/issues/44650
Fixes: https://github.com/nodejs/node/issues/37472

+++ deps/v8/src/bigint/tostring.cc
@@ -600,7 +600,8 @@ int ToStringResultLength(Digits X, int radix, bool sign) {
     uint64_t chars_required = bit_length;
     chars_required *= kBitsPerCharTableMultiplier;
     chars_required = DIV_CEIL(chars_required, min_bits_per_char);
-    DCHECK(chars_required < std::numeric_limits<int>::max());
+    DCHECK(chars_required <
+           static_cast<uint64_t>(std::numeric_limits<int>::max()));
     result = static_cast<int>(chars_required);
   }
   result += sign;
